# Otros/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/vacaciones/{id}":{"get":{"tags":["vacaciones-controller"],"operationId":"getMatriculaById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericResponseVacacionesDTO"}}}}}},"put":{"tags":["vacaciones-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VacacionesDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericResponseVacacionesDTO"}}}}}},"delete":{"tags":["vacaciones-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/vacaciones":{"get":{"tags":["vacaciones-controller"],"operationId":"getallMatriculas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericResponseVacacionesDTO"}}}}}},"post":{"tags":["vacaciones-controller"],"operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VacacionesDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"PersonaDTO":{"required":["apellidosMaterno","apellidosPaterno","dni","edad","nombresPersona"],"type":"object","properties":{"vacaciones":{"$ref":"#/components/schemas/VacacionesDTO"},"nombresPersona":{"maxLength":60,"minLength":1,"type":"string"},"apellidosPaterno":{"type":"string"},"apellidosMaterno":{"type":"string"},"dni":{"type":"string"},"edad":{"type":"integer","format":"int32"}}},"VacacionesDTO":{"required":["descripcion","detallePersona","estado","fechaFin","fechaInicio"],"type":"object","properties":{"idVacaciones":{"type":"integer","format":"int32"},"descripcion":{"maxLength":50,"minLength":5,"type":"string"},"fechaInicio":{"type":"string","format":"date-time"},"fechaFin":{"type":"string","format":"date-time"},"estado":{"type":"boolean"},"detallePersona":{"type":"array","items":{"$ref":"#/components/schemas/PersonaDTO"}}}},"GenericResponseVacacionesDTO":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/VacacionesDTO"}}}}}}}